% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expression_get}
\alias{expression_get}
\title{Get GEO data}
\usage{
expression_get(gse, ...)
}
\arguments{
\item{gse}{A character vector of the GEO series IDs}

\item{...}{Other arguments passed to getGEO}
}
\description{
Get GEO data
}
\examples{
\dontrun{
gse <- c('GSE34150', 'GSE15018', 'GSE20696', 'GSE69313')
expression_get(gse, destdir = 'data')
}

}
