% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expression_subset}
\alias{expression_subset}
\title{Subset an ExpressionSet}
\usage{
expression_subset(eset, byrow, bycol, collapse_rows = FALSE, ...)
}
\arguments{
\item{eset}{An ExpressionSet object}

\item{byrow}{A logical vector masking rows of eset}

\item{bycol}{A logical vector masking columns of eset}

\item{collapse_rows}{A logical (default FALSE)}

\item{...}{Other arguments passed to collapse_rows}
}
\value{
A tidy expression matrix. The dimensions of the matrix are either
equal to the byrow and bycol argumnets when collapse_rows is FALSE, or equal
the rowGroup and the bycol arguments otherwise
}
\description{
Subset an ExpressionSet object by row and/or column, and perform collapse of
repeated probes corresponding to the same gene ID
}
\examples{
library(Biobase)
data("sample.ExpressionSet")

pd <- pData(sample.ExpressionSet)

set.seed(123)
byrow <- sample(c(TRUE, FALSE), nrow(sample.ExpressionSet), replace = TRUE)
bycol <- pd$sex == 'Female'

expression_subset(sample.ExpressionSet, byrow, bycol)

}
