% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{interactions_get}
\alias{interactions_get}
\title{Get STRING interactions}
\usage{
interactions_get(genes, ...)
}
\arguments{
\item{genes}{A data.frame of at least one column of gene sybols}

\item{...}{Other arguments passed to new}
}
\value{
A data.frame
}
\description{
Get STRING interactions
}
